+++ lib/shark/permission.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Shark
+  class Permission < Base
+    extend ContactService::Resource
+  end
+end